01_etl_pipeline/requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo>=4.0

-e .